Switch to np.squeeze() for data matrices
Getting a single cell or a single trial from a `Fluorescence` or
`TrialDataset` yields an array with one or more dimensions of size one.
For example,

>>> fluorescence.get_cells(0).data.shape
[10, 1, 462]

The resulting array can usually be expressed as a matrix, which is
convenient for passing to matplotlib plotting functions. Before this
patch, this was done by manual slicing to remove dimensions of length
one.

>>> plt.plot(fluorescence.get_cells(0).data[:, 0, :].T)

This patch switches to using `np.squeeze()`, which produces the same
result without having to remember the order of the dimensions (except
that time is last).

>>> plt.plot(fluorescence.get_cells(0).data.squeeze().T)